UI Controller Expansion
We're starting to have a lot of strings with the querySelector which is a problem. Imagine we make changes to the class names in the UI, then we would have to change the strings manually which could be expensive. Lets create an object where we store all of that data and since we're talking about the UI, it should be done in the UI Controller.